requirements.txt
# Google APIs
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2

# Document Parsing
PyPDF2==3.0.1
python-docx==1.1.0
olefile==0.47
pyhwp==0.1b12
openpyxl==3.1.2
# pywin32==306  # Legacy .doc file support (Windows only - not compatible with Linux/Cloud Run)

# Vector Store
chromadb>=0.4.22
numpy<2.0.0  # ChromaDB compatibility

# Text Processing (LangChain 제거 - 의존성 충돌 방지)
# langchain>=0.1.0
# langchain-text-splitters>=0.0.1
# langchain-community>=0.0.38
# tiktoken>=0.5.0

# API Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
